backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.11.0
pydantic==2.9.2
python-dotenv==1.0.1
openai==2.28.0
requests==2.32.3
google-genai==1.67.0
